Add Redis session manager
This session manager stores sessions in a redis database.
It allows for storing any kind of value in the session assuming that
their exists a json formatter for the session objects. It is based on
the spray-json module.
Session timeout relies on the expiration feature of Redis for keys, and
thus no cleanup code is necessary on the manager side.